fix(build): Fix APIENTRY redefinition warnings, printf format casts, and add /FS MSVC option

+++ src/client/renderer/Tesselator.cpp
@@ -39,7 +39,7 @@ Tesselator::Tesselator( int size )
 
 	char* a = (char*)&_varray[0];
 	char* b = (char*)&_varray[1];
-	LOGI("Vsize: %lu, %d\n", sizeof(VERTEX), (b-a));
+	LOGI("Vsize: %lu, %d\n", (unsigned long)sizeof(VERTEX), (int)(b - a));
 }
 
 Tesselator::~Tesselator()
